requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.1.0
numpy>=1.26.0
gspread>=5.12.0
google-auth>=2.23.0
python-dotenv>=1.0.0
tenacity>=8.2.0
PyYAML>=6.0.1
fake-useragent>=1.4.0
